New LLM primitive
