+++ packages/core/package.json
@@ -78,6 +78,7 @@
     "@opentelemetry/sdk-trace-node": "^1.28.0",
     "@opentelemetry/semantic-conventions": "^1.28.0",
     "@pinecone-database/pinecone": "^3.0.3",
+    "@portkey-ai/vercel-provider": "^1.0.1",
     "@types/express": "^5.0.0",
     "@upstash/redis": "^1.34.3",
     "ai": "^4.0.34",
@@ -87,6 +88,7 @@
     "dotenv": "^16.4.7",
     "express": "^4.21.1",
     "json-schema": "^0.4.0",
+    "ollama-ai-provider": "^1.2.0",
     "pino": "^9.6.0",
     "pino-pretty": "^13.0.0",
     "radash": "^12.1.0",
