+++ packages/core/src/llm/aisdk/providers/cohere.ts
@@ -2,13 +2,15 @@ import { createCohere } from '@ai-sdk/cohere';
 
 import { AISDK } from '../aisdk';
 
+export type CohereModel = 'command-r-plus' | (string & {});
+
 export class Cohere extends AISDK {
   constructor({
     name = 'command-r-plus',
     apiKey = process.env.COHERE_API_KEY || '',
     baseURL = 'https://api.cohere.com/v2',
   }: {
-    name?: string;
+    name?: CohereModel;
     apiKey?: string;
     baseURL?: string;
   }) {
