+++ packages/core/src/llm/aisdk/providers/perplexity.ts
@@ -2,17 +2,7 @@ import { createPerplexity } from '@ai-sdk/perplexity';
 
 import { AISDK } from '../aisdk';
 
-export type PerplexityModel =
-  | 'llama-3.1-sonar-small-128k-online'
-  | 'llama-3.1-sonar-large-128k-online'
-  | 'llama-3.1-sonar-huge-128k-online'
-  | 'llama-3.1-sonar-small-128k-chat'
-  | 'llama-3.1-sonar-large-128k-chat'
-  | 'llama-3.1-8b-instruct'
-  | 'llama-3.1-70b-instruct'
-  | 'sonar'
-  | 'sonar-pro'
-  | (string & {});
+export type PerplexityModel = 'sonar' | 'sonar-pro' | (string & {});
 
 export class Perplexity extends AISDK {
   constructor({
